schema changes to faciltiate dynamo

+++ backend/entities/plan.py
@@ -31,7 +31,7 @@ class PlanMilestone(BaseModel):
     date: str
     description: str
     criteria: Optional[List[Union[PlanMilestoneCriteria, PlanMilestoneCriteriaGroup]]] = Field(default=None, description="The criteria that need to be met to achieve the milestone. If unexistent it means the milestone progress is manually updated.")
-    progress: Optional[float] = None  # Progress as a percentage (0-100)
+    progress: Optional[int] = None  # Progress as a percentage (0-100)
 
 
 class Plan(BaseModel):
